+++ examples/blueprints2-example/imgui_extras.cpp
@@ -1,8 +1,8 @@
-# include "drawing.h"
+# include "imgui_extras.h"
 # define IMGUI_DEFINE_MATH_OPERATORS
 # include <imgui_internal.h>
 
-void ax::Drawing::DrawIcon(ImDrawList* drawList, const ImVec2& a, const ImVec2& b, IconType type, bool filled, ImU32 color, ImU32 innerColor)
+void ImEx::DrawIcon(ImDrawList* drawList, const ImVec2& a, const ImVec2& b, IconType type, bool filled, ImU32 color, ImU32 innerColor)
 {
           auto rect           = ImRect(a, b);
           auto rect_x         = rect.Min.x;
@@ -220,3 +220,25 @@ void ax::Drawing::DrawIcon(ImDrawList* drawList, const ImVec2& a, const ImVec2& 
         }
     }
 }
+
+
+void ImEx::Icon(const ImVec2& size, IconType type, bool filled, const ImVec4& color/* = ImVec4(1, 1, 1, 1)*/, const ImVec4& innerColor/* = ImVec4(0, 0, 0, 0)*/)
+{
+    if (ImGui::IsRectVisible(size))
+    {
+        auto cursorPos = ImGui::GetCursorScreenPos();
+        auto drawList  = ImGui::GetWindowDrawList();
+        DrawIcon(drawList, cursorPos, cursorPos + size, type, filled, ImColor(color), ImColor(innerColor));
+    }
+
+    ImGui::Dummy(size);
+}
+
+void ImEx::Debug_DrawItemRect(const ImVec4& col)
+{
+    auto drawList = ImGui::GetWindowDrawList();
+    auto itemMin = ImGui::GetItemRectMin();
+    auto itemMax = ImGui::GetItemRectMax();
+    drawList->AddRect(itemMin, itemMax, ImColor(col));
+}
+
